move label logic to backend
