+++ front/lib/resources/agent_mcp_action_resource.ts
@@ -1038,6 +1038,7 @@ export class AgentMCPActionResource extends BaseResource<AgentMCPActionModel> {
       internalMCPServerName,
       mcpServerName: this.toolConfiguration.mcpServerName,
       toolName,
+      inputs: this.augmentedInputs,
     });
   }
